Break out the `cancel` command
Continuing with the clean-up of `pkg/cmd/pulumi`, this factors out the
`cancel` command.
