+++ pkg/cmd/pulumi/cancel/cancel.go
@@ -1,4 +1,4 @@
-// Copyright 2016-2023, Pulumi Corporation.
+// Copyright 2016-2024, Pulumi Corporation.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package main
+package cancel
 
 import (
 	"errors"
@@ -32,7 +32,7 @@ import (
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/result"
 )
 
-func newCancelCmd() *cobra.Command {
+func NewCancelCmd() *cobra.Command {
 	var yes bool
 	var stack string
 	cmd := &cobra.Command{
